+++ next.config.js
@@ -52,7 +52,7 @@ const nextConfig = {
   reactStrictMode: true,
   trailingSlash: true,
   env: {
-    version: '3.4.0'
+    version: '3.5.0'
   },
   devIndicators: {
     buildActivity: false